chore(backend): rename /src/info to application ifo and added tests for DB connected/no connected
