+++ backend/src/applicationInfo/version.json
